requirements copy.txt
Pillow==4.0.0
watchdog==0.8.3